updated package.json
Updated package.json to force TypeScript 5 to work with current version of react-scripts. This is because the developer of react-scripts have yet to release a new version to be compatible with TS 5.

Also added gh-pages package, and set up everything to be ready for GitHub Pages deployment, which is to be deployed by running 'npm run deploy'

Subsequent changes to the main branch, when ready for deployment, have to be deployed by calling 'npm run deploy' again